requirements.txt
fastapi
uvicorn
faiss-cpu
transformers
sentence-transformers
scikit-learn
